client online work
implemented:
 - screens for all asset types - memo, cred, card, file
 - create/display remote  functions

TODO - implement update
